requirements.txt
Flask==1.1.1
gunicorn==20.0.4
Flask-SQLAlchemy==2.4.1
psycopg2-binary==2.8.4
python-dotenv == 0.10.3
pytz == 2020.1
python-dateutil == 2.8.1
Flask-RESTful == 0.3.8
flask-restplus == 0.13.0





